werewolves/requirements.txt
Flask==0.12.1
gunicorn==19.7.1
google-cloud-datastore==1.0.0
flask-cors==3.0.2